Update README.md
In this project I have used the Convolutional Neural Network for the classification
of the wafermap images. In particular the goal was to correct classifying the various
types of error that can occur in the production of the wafer(semiconductor). They
have been used different techniques to overcome problems related to:images of different size,
unbalanced classes, useful metrics.